separating gulp into multiple files to improve reading
